main: Block SIGUSR1 instead of ignoring it
wlroots uses SIGUSR1 internally to signal that xwayland started up.  The
signal is received in the wayland event loop (due to
wl_event_loop_add_signal()) so all other threads need to block that
signal or it might end up in the wrong thread.

The code comment was correct but code wise we ignored the signal instead
of blocking it. This caused the signal to sometimes be sent to the main
thread instead of the one running the wayland event loop which listens
for SIGUSR1.

Closes: #20