input method: Forward only supported state
The input-method state is a reflection of the text-input state. Sending all availble state change events to the input-method on each done doesn't actually reflect the state text-input is in, if the values were filled in from defaults because text-input doesn't support them.

With this change, support is taken into account, and defaulted values are not sent at all.