render: Rework damage tracking debug mode
Before this change, the damage tracking debug flag enabled a mode
that displayed only the updated contents of the screen and filled
the rest with solid color. This may be useful when developing
wlroots' damage tracking, but it's not very effective for practically
anything else, since actually using the device in such mode is
close to impossible.

This commit changes it in a way that makes it mark the updated
screen areas with semi-transparent color rectangles. Yellow color
represents currently updated area, while magenta shows areas that
have been updated in a last frame (so single frame changes can still
be visually inspected).