src: add utility function to fix rotation on wlroots v0.11+
Starting with version 0.11.0, wlroots rotates counter-clockwise, while
it rotated clockwise in previous version.

In order to keep the same behavior no matter which wlroots version is
used, this patch adds a new `phoc_utils_fix_transform()` function and
uses it for correcting rotation requests.

This makes sure we always rotate clockwise, as it was the previous
behavior.