Remove view rotation handling
It's not really a useful feature to have in a mobile compositor and just
makes positioning code more complicated for no good reason. It wasn't
even complete, as evidenced by TODOs being removed with it. Support for
iterating over and drawing rotated surfaces is retained as it may still
be useful, but we sure don't need it at the view level.