phosh-private: Allow clients to be informed about application startup
This allows interested clients to be notified about an apps startup
sequence. We add two events: One is used whan the app is launched
(spawned) and the other one when the spawned client send it's
startup-id.

This is modeled after gtk_shell1 wayland protocol but can be extended to
cover other protocols like xdg-activation or X startup-notification spec
in the future.