Thumbnail rendering with wlr_allocator
This part of wlroots API is not public at the moment, so we copy the headers.
We're already using unstable API interfaces, so we don't really lose anything
by doing that. wlr_allocator is supposed to become public at some point, but
it's not known yet when and in which form.