settings: Drop switch configuration
E.g. g-s-d uses logind and upower for this so we don't need it
in the compositor atm.

We keep the switch code since we might want to forward this to the shell
at some point (e.g. for docked detection of convertibles).

This leaves only the output configuration and xwayland enablement in
phoc.ini which is useful since we don't have a proper replacement yet.