server: Don't assert when we fail to initialize the backend
So far we asserted which confuses users thinking that there's someting
wrong in phoc when usually there's something wrong with their EGL or DRM
setup.

The solution is not ideal since the servier is a singleton and hence
we need to evaulate the error in `phoc_server_get_default ()` but since
it stays around forever that doesn't hurt much atm.