view: Add a GObject as first element
The two implementations of the PhocView interface (roots_xdg_surface and
roots_xwayland_surface) expect PhocView to be the first element in their
structs but since we also need GObject as first element to introduce
PhocXdgView and PhocXWaylandView (which in turn need GObject as first
struct member for tat) we need to put a GObject as first struct member
into PhocView.

This will fix itself once PhocXdgView, PhocXWaylandView and PhocView are
GObjects and PhocXdgView and PhocXWayland derive from PhocView. Until
then it's an interim hack.