output: Invoke arrange-layers less
phoc_layer_shell_arrange() needs to be called on mode, scale and
transform changes to adjust the layer surfaces and available space for
views. It doesn't need to be called when we only submit damage or attach
new buffers. This reduces the number of invocations (and hence used CPU
time).

Fixes: 3d788792 ("Update to wlroots 0.13 and drop support for older versions")
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/352>