Adjust to render changes
The output now needs wlr_output_init_render called and
wlr_backend_get_renderer is gone but since we can now get the renderer
from the output changes initially aren't that big.

Since most of the functions can fail we need to shuffle some things
around but that happens in the next commits.