output: Use a single list for all layer surfaces
This has the advantage that we can use a single loop in most cases and
moving surfaces between layers keeps the ordering intact.

Performance wise it doesn't make much of a difference as the lists are
small.

This also simplifies the API as we pass output and wanted layer around
rather than per layer lists.