output: Automatically compute scaling factor
`phoc` currently defaults to unity scaling, unless the corresponding
output configuration contains a `scale` value. Whenever the physical
size of the output is available, we could actually calculate this value
in order to reach the best apparent DPI. This will allow distributors to
get rid of device-specific config files.

The scaling factor is automatically calculated as long as the `scale`
option is absent from the output configuration, defaulting to `1` if the
value can't be calculated (for example, a virtual output reporting a
physical size of 0 x 0 mm).

In addition, this commit allows setting `scale = auto` in order to let
users explicitly specify the scalei value should be automatically
calculated.