xdg-shell: Use toplevels output when unconstraining popups
This fixes popups being placed on the wrong output when close to
a screen edge when there's another outut nearby.

This can e.g. be seen when using gnome-calendar maximized with another
output to the right. In this case wlr_output_layout_closest_point()
would return the closer output to the right instead of the toplevel
ones.