Add timed animations
The PhocTimedAnimation can be used to animate properties of Objects by
specifying a PhocPropertyEaser that adjusts the properties over the
course of the animation. The PhocPropertyEaser can handle int and float
properties.

The Objects must implement the PhocAnimatable interface so the animation
can be driven by the Objects preferred frame clock.