layer-shell-effects: Calculate slide duration based on distance to travel
Previously phoc would use the same duration for slide animation regardless
of how far the surface had to slide, which caused it to be unnaturally slow
in cases when the surface had to slide for a very short distance. Fix it
by shortening the animation duration based on how far the surface has been
already (un)folded at the beginning of animation.