layer-shell-effects: Make sure to apply new margin on animation end
We were forgetting to set the margin and trigger layer arrangement
at the very last frame of the animation. Usually this wouldn't
matter since the last frame doesn't usually change the margin,
but this may matter in cases where we need to correct for
overshooting, which may happen if the surface has changed its
target margins.