layer-shell-effects: Don't assert on NULL layer-surface
Don't abort the compositor when the client already destroyed the layer
surface but not yet the drag surface.

Except for phoc_draggable_layer_surface_slide the functions
don't actually access the layer-surface but it's easier to be
consistent in case we need to access it later on when adding new
features.