idle-inhibit: Use app-id of inhibiting view if possible
If we have an app-id for the view that inhibits idle use it's app-id
when requesting the inhibitor.

This makes sure applications listing inhibitors can print information
about the inhibiting application. This is e.g. the case for phosh's
end-session-dialog.

Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/446>