keyboard: Forward on the super key if pressed by itself
If the super key is the only key pressed (so it isn't used as a
modifier) we want to pass the event onto Phosh. This handles the complex
logic of looking for the key being pressed and released in Phoc allowing
Phosh to just easily react to the event.

Fixes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/563
Signed-off-by: Alistair Francis <alistair@alistair23.me>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/447>