view: Implement phoc_view_child_damage_whole for popups
So far we damaged the parent surface which doesn't work well when the
popup extends out of the parent. It also avoids that we submit way
more damage than necessary.

We ought to implement `get_pos` for subsurfaces as well at some point.

Closes: https://gitlab.gnome.org/World/Phosh/phoc/-/issues/325
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/465>