ci: Move style checks to separate stage
Style checks often fail due to minor nits. When looking at the pipeline
state we want to see at a single glance if this is just a style check or
maybe an asan failure.

So make sure these aren't in the same stage.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/500>