seat: Avoid misplacing when dragging maximized views
If a toplevel is maximized initially there's no saved width or height.
Use the maximized width/height in this case. Otherwise the toplevel's top
left corner suddenly jumps to the cursor position when dragging this
toplevel via mouse pointer from it's maximized state.

This can be reproduced with firefox. Use

   WLR_BACKENDS=wayland   _build/src/phoc -E firefox

and maximize firefox. Then close firefox. This will make it restore
the maximized state on next startup:

   WLR_BACKENDS=wayland   _build/src/phoc -E firefox

Now drag the window on the titlebar and notice the top left corner
jumping to the cursor position. With this patch this will no longer
happen.

Closes: https://gitlab.gnome.org/World/Phosh/phoc/-/issues/349

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/503>