output: Mostly move damage ring handling to output
Even after moving the full screen view scan out and the frame-done
event handling the current render pass is a mixed bag of rendering,
adding additional damage for debug options (touch-points, damage) and
comittng the new output state.

As a next step move the rotation of the damage ring and the output
commit out of the render_output call.

This makes the render flow easier to track. We pass the render context
around for that and hence make it public. Context is a slightly better
fit than the generic data.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/510>