view: Take geometry into accout for fullscreen views
If we hit the direct scanout path geometry doesn't matter for the
rendered surface but when the user e.g. opens a popup we hit the
rendering path and we need to take geometry into account as otherwise the
surface is shifted towards the bottom right edge.

This also fixes UI elements not being responsive in fullscreen.

Closes: https://gitlab.gnome.org/World/Phosh/phoc/-/issues/356
Closes: https://gitlab.gnome.org/World/Phosh/phoc/-/issues/357

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/520>