desktop: Handle always-on-top
This allows views to stay on top (e.g. video players).

When a view is raised it's already mapped as well as new children are
raised too to avoid a parent hiding it's children. We don't raise any
parents in order to e.g. allow to raise a single window of an
application.

To activate you need to bind a keyboard shortcut

  gsettings set org.gnome.desktop.wm.keybindings always-on-top "['<super>t']"

Closes: https://gitlab.gnome.org/World/Phosh/phoc/-/issues/342
Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/536>