view: Make view-child a GObject
We use the same pattern as with PhocView and allow for public struct
members so derived classes can access them easily. Since PhocView and
PhocViewChild are still entangled we leave most of the member functions
in PhocView for the moment. They'll be moved in follow up commits.

We need to rename phoc_view_child_init() to not clash with the GObject
init function.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/548>