seat: Reverse window cycle focus order
Cycling "backward" now means to focus the view that previously had focus
while moving forward means to focus the view that was out of focus for
the longest time.

With this we correctly focus the last focused view when cycling via
keyboard shortcuts and then closing a window.

It also fixes the case where opening an application and cycling backward
would not get one back to the previous application.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/558>