view: Don't draw decoration on maximized windows
No need to do that and it helps to not leak decorations into e.g.
transparent top or bottom layer surfaces.

Later on we should restrict this to auto-maxmize mode and rather
calculate the maximized window size *including* decorations so that
they're visible and usable in non auto-maximized mode.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/570>