layer-shell-effects: When folded reject fling that wouldn't unfold
When a surface is folded we reject a fling that doesn't go into the
unfold direction.

This prevents us picking up the gesture when an ongoing fling on another
surface crosses this surface. This can e.g. happen when swipe closing
apps in phosh where one crosses the top bar during the fling.

Closes: https://gitlab.gnome.org/World/Phosh/phoc/-/issues/380
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/578>