layer-shell: Make child creation/destruction available to PhocLayerSurface
We want to move the PhocLayerSurface handling to the class
implementation. Since popups and subsurfaces are currently handled in
layer-shell we need to make them available to PhocLayerSurface.

We do so by introducing a private header which we can then reuse when we
move the popups and subsurfaces out of layer-shell too.

This leaks a bit of popup setup details as we expose
phoc_layer_popup_unconstrain() too but we can fix that up when
introducing the PhocLayerPopup object.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/582>