layer-surface: Invalidate layer surface lists less often
It is sufficient if we invalidate the list when the surfaces changes
layers or exclusive zone (which can affect the ordering within a layer).
This makes us hit the fast path in
  `phoc_output_get_layer_surfaces_for_layer()``
way more often as we don't invalidate on every commit.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/616>