layer-surface: Notify transaction about configure events
We notify the transaction system when we send a `configure` event to a
layer surface (potentially kicking off a new transaction). This will
later on allow us to render the current buffers while the transaction
is ongoing.

When a layer surface client commits a new buffer we check if the
corresponding ack'ed configure serial is recent enough and if so,
inform the transaction system that we have a recent enough buffer
for this layer-surface.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/619>