child-root: New interface that is the root of a ViewChild tree
We want either a layer surface or a view to be the root so introduce an
interface. We call it `PhocRootChild` to not interfere with other
possible roots (e.g. the root of a view tree).

As a follow up we want to rename the `PhocViewChild` to just
`PhocChild`.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/627>