main: Add --verbose option
wlroots uses `info` for informative (non-debug) messages, info log level
isn't printed by glib though, so allow to use `message` for it.

We often ask ourselves: Which renderer and which backend was in use?
Which GL extensions were detected. wlroots prints this on startup but it
is very noisy, so add an option `-v` we can add to `phosh-session` but
not use when people use phoc for nested testing, etc.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/641>