cairo-texture: Introduce PhocCairoTexture
This object wraps up the bits necessary to create a wlr_texture that is
backed by a Cairo image surface. It's adapted from the wlroots
cairo-buffer example code.

It will be used in the next commit to handle rendering for PhocSpinner.

Signed-off-by: Sam Day <me@samcday.com>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/646>