xdg-popup: Ignore reposition for layer-shell
GTK4 likes to reposition popups however for layer-shell this doesn't
work as wlr_xdg_popup_get_position is tied to xdg-surface and asserts if
passed in a layer-surface. Skip the reposition in this case.

Another approach would be to not connect the signal in the first place
when handling `new_popup`.

This fixes a crash with cameractl (https://gitlab.com/NekoCWD/cameractrl.git)

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/663>