idle-inhibit: Hold reference while taking inhibitor
If an inhibit is followed by an uninhibit the inhibit call might
get cancelled, we thus never get the cookie and thus can't clean
up the inhibitor. Fix this by not cancelling inhibotor calls but
rater taking a reference while getting the inhibitor is in flight.

The simplest thing would be to use sync DBus calls but that could block
the compositor.

Closes: https://gitlab.gnome.org/World/Phosh/phoc/-/issues/398

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/668>