desktop: Don't rebuild output configs for state changes
Now that `phoc_output_config_head_to_output_config` builds an
output config for filling in the pending state we can use that
same config to persist the outputs states.

This makes sure we don't have yet another place that needs to
keep track of the output configuration. The only thing the
desktop needs to care about is to get the identifier right.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/690>