testlib: Store format in GPtrArray
The current bitfield isn't large enough for large format values
which doesn't matter in practice (as we're only interested on
small valued) but it trips up clang.

We use a GPtrArray instead of a GArray due to the easier searching.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/695>